+++ packages/core/src/charts/combo.ts
@@ -35,14 +35,18 @@ const graphComponentsMap = {
 	[ChartTypes.LINE]: [Line, Scatter],
 	[ChartTypes.SCATTER]: [Scatter],
 	[ChartTypes.AREA]: [Area, Line, Scatter],
-	[ChartTypes.STACKED_AREA]: [StackedArea, Line, StackedScatter, StackedRuler],
+	[ChartTypes.STACKED_AREA]: [
+		StackedArea,
+		Line,
+		StackedScatter,
+		StackedRuler
+	],
 	[ChartTypes.SIMPLE_BAR]: [SimpleBar],
 	[ChartTypes.GROUPED_BAR]: [GroupedBar, ZeroLine],
 	[ChartTypes.STACKED_BAR]: [StackedBar, StackedRuler]
 };
 
 export class ComboChart extends AxisChart {
-
 	constructor(holder: Element, chartConfigs: ChartConfig<ComboChartOptions>) {
 		super(holder, chartConfigs);
 
@@ -58,7 +62,9 @@ export class ComboChart extends AxisChart {
 		if (!chartConfigs.options.comboChartTypes) {
 			console.error("No comboChartTypes defined for the Combo Chart!");
 			// add a default chart to get an empty chart
-			chartOptions.comboChartTypes = [{ type: ChartTypes.LINE, correspondingDatasets: [] }];
+			chartOptions.comboChartTypes = [
+				{ type: ChartTypes.LINE, correspondingDatasets: [] }
+			];
 		}
 
 		// set the global options
@@ -71,32 +77,58 @@ export class ComboChart extends AxisChart {
 	getGraphComponents() {
 		const { comboChartTypes } = this.model.getOptions();
 		let counter = 0;
-		const graphComponents = comboChartTypes.map((graph) => {
-			const type = graph.type;
-			let options;
+		const graphComponents = comboChartTypes
+			.map((graph) => {
+				const type = graph.type;
+				let options;
 
-			// initializes the components using input strings with the base configs for each chart
-			if (typeof graph.type === "string") {
-				// check if it is in the components map
-				// if it isn't then it is not a valid carbon chart to use in combo
-				if (!Object.keys(graphComponentsMap).includes(graph.type)) {
-					console.error(`Invalid chart type "${graph.type}" specified for combo chart. Please refer to the ComboChart tutorial for more guidance.`);
-					return null;
+				// initializes the components using input strings with the base configs for each chart
+				if (typeof graph.type === "string") {
+					// check if it is in the components map
+					// if it isn't then it is not a valid carbon chart to use in combo
+					if (!Object.keys(graphComponentsMap).includes(graph.type)) {
+						console.error(
+							`Invalid chart type "${graph.type}" specified for combo chart. Please refer to the ComboChart tutorial for more guidance.`
+						);
+						return null;
+					}
+					let stacked;
+					options = Tools.merge(
+						{},
+						Configuration.options[
+							`${Tools.camelCase(graph.type)}Chart`
+						],
+						this.model.getOptions(),
+						graph.options
+					);
+					// if we are creating a stacked area, the contained Line chart needs to know it is stacked
+					if (graph.type === ChartTypes.STACKED_AREA) {
+						stacked = true;
+					}
+					return graphComponentsMap[graph.type].map(
+						(Component, i) =>
+							new Component(this.model, this.services, {
+								groups: graph.correspondingDatasets,
+								id: counter++,
+								options: options,
+								stacked
+							})
+					);
+				} else {
+					// user has imported a type or custom component to instantiate
+					options = Tools.merge(
+						{},
+						this.model.getOptions(),
+						graph.options
+					);
+					return new type(this.model, this.services, {
+						groups: graph.correspondingDatasets,
+						id: counter++,
+						options: options
+					});
 				}
-				let stacked;
-				options = Tools.merge({}, Configuration.options[`${Tools.camelCase(graph.type)}Chart`], this.model.getOptions(), graph.options);
-				// if we are creating a stacked area, the contained Line chart needs to know it is stacked
-				if (graph.type === ChartTypes.STACKED_AREA) {
-					stacked = true;
-				}
-				return graphComponentsMap[graph.type].map((Component, i) =>
-					new Component(this.model, this.services, { groups: graph.correspondingDatasets, id: counter++, options: options, stacked }));
-			} else {
-				// user has imported a type or custom component to instantiate
-				options = Tools.merge({}, this.model.getOptions(), graph.options);
-				return new type(this.model, this.services, { groups: graph.correspondingDatasets, id: counter++, options: options });
-			}
-		}).filter(item => item !== null);
+			})
+			.filter((item) => item !== null);
 
 		return Tools.flatten(graphComponents);
 	}
@@ -104,8 +136,11 @@ export class ComboChart extends AxisChart {
 	getComponents() {
 		const { comboChartTypes } = this.model.getOptions();
 		// don't add the regular ruler if stacked ruler is added
-		const stackedRulerEnabled = comboChartTypes.some(chartObject =>
-			chartObject.type === ChartTypes.STACKED_BAR || chartObject.type === ChartTypes.STACKED_AREA);
+		const stackedRulerEnabled = comboChartTypes.some(
+			(chartObject) =>
+				chartObject.type === ChartTypes.STACKED_BAR ||
+				chartObject.type === ChartTypes.STACKED_AREA
+		);
 
 		// Specify what to render inside the graph-frame
 		const graphFrameComponents = [
@@ -114,11 +149,15 @@ export class ComboChart extends AxisChart {
 			new Skeleton(this.model, this.services, {
 				skeleton: Skeletons.GRID
 			}),
-			...(stackedRulerEnabled ? [] : [new Ruler(this.model, this.services)]),
+			...(stackedRulerEnabled
+				? []
+				: [new Ruler(this.model, this.services)]),
 			...this.getGraphComponents()
 		];
 
-		const components: any[] = this.getAxisChartComponents(graphFrameComponents);
+		const components: any[] = this.getAxisChartComponents(
+			graphFrameComponents
+		);
 
 		return components;
 	}