+++ packages/core/src/components/graphs/line.ts
@@ -46,7 +46,9 @@ export class Line extends Component {
 			.y(getYValue)
 			.curve(curves.getD3Curve())
 			.defined((datum: any, i) => {
-				const rangeIdentifier = cartesianScales.getRangeIdentifier(datum);
+				const rangeIdentifier = cartesianScales.getRangeIdentifier(
+					datum
+				);
 				const value = datum[rangeIdentifier];
 				if (value === null || value === undefined) {
 					return false;
@@ -60,12 +62,19 @@ export class Line extends Component {
 				(axis) => options.axes[axis].percentage
 			);
 			const { groupMapsTo } = options.data;
-			const stackedData = this.model.getStackedData({groups: this.configs.groups, percentage });
+			const stackedData = this.model.getStackedData({
+				groups: this.configs.groups,
+				percentage
+			});
 
 			data = stackedData.map((d) => {
-				const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(d);
-				const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(d);
-				return ({
+				const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(
+					d
+				);
+				const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(
+					d
+				);
+				return {
 					name: d[0][groupMapsTo],
 					data: d.map((datum) => ({
 						[domainIdentifier]: datum.data.sharedStackKey,
@@ -73,7 +82,7 @@ export class Line extends Component {
 						[rangeIdentifier]: datum[1]
 					})),
 					hidden: !Tools.some(d, (datum) => datum[0] !== datum[1])
-				});
+				};
 			});
 		} else {
 			data = this.model.getGroupedData(this.configs.groups);
@@ -108,16 +117,20 @@ export class Line extends Component {
 					originalClassName: "line"
 				})
 			)
+			.style("stroke", (group) => this.model.getStrokeColor(group.name))
 			// a11y
 			.attr("role", Roles.GRAPHICS_SYMBOL)
 			.attr("aria-roledescription", "line")
 			.attr("aria-label", (group) => {
 				const { data: groupData } = group;
 				return groupData
-					.map((datum) =>  {
-						const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(datum);
+					.map((datum) => {
+						const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(
+							datum
+						);
 						return datum[rangeIdentifier];
-					}).join(",");
+					})
+					.join(",");
 			})
 			// Transition
 			.transition(
@@ -127,7 +140,6 @@ export class Line extends Component {
 				)
 			)
 			.attr("opacity", (d) => (d.hidden ? 0 : 1))
-			.attr("stroke", (group) => this.model.getStrokeColor(group.name))
 			.attr("d", (group) => {
 				const { data: groupData } = group;
 				return lineGenerator(groupData);
@@ -149,7 +161,7 @@ export class Line extends Component {
 
 				return Configuration.lines.opacity.selected;
 			});
-	}
+	};
 
 	handleLegendMouseOut = (event: CustomEvent) => {
 		this.parent
@@ -158,7 +170,7 @@ export class Line extends Component {
 				this.services.transitions.getTransition("legend-mouseout-line")
 			)
 			.attr("opacity", Configuration.lines.opacity.selected);
-	}
+	};
 
 	destroy() {
 		// Remove event listeners