+++ packages/core/src/components/graphs/scatter.ts
@@ -8,6 +8,7 @@ import { select, Selection } from "d3-selection";
 
 export class Scatter extends Component {
 	type = "scatter";
+	scatterData: any;
 
 	init() {
 		const { events } = this.services;
@@ -38,7 +39,9 @@ export class Scatter extends Component {
 	}
 
 	filterBasedOnZoomDomain(data) {
-		const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(data);
+		const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(
+			data
+		);
 		const zoomDomain = this.model.get("zoomDomain");
 		if (zoomDomain !== undefined) {
 			return data.filter(
@@ -48,6 +51,38 @@ export class Scatter extends Component {
 			);
 		}
 		return data;
+	}
+
+	getScatterData() {
+		const options = this.getOptions();
+
+		const { stacked } = this.configs;
+
+		let scatterData;
+		if (stacked) {
+			const percentage = Object.keys(options.axes).some(
+				(axis) => options.axes[axis].percentage
+			);
+			scatterData = this.model.getStackedData({
+				groups: this.configs.groups,
+				percentage
+			});
+		} else {
+			scatterData = this.model
+				.getDisplayData(this.configs.groups)
+				.filter((d) => {
+					const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(
+						d
+					);
+					return (
+						d[rangeIdentifier] !== undefined &&
+						d[rangeIdentifier] !== null
+					);
+				});
+		}
+
+		// filter out datapoints that aren't part of the zoomed domain
+		return this.filterBasedOnZoomDomain(scatterData);
 	}
 
 	render(animate: boolean) {
@@ -65,31 +100,15 @@ export class Scatter extends Component {
 		const options = this.getOptions();
 		const { groupMapsTo } = options.data;
 
-
-		const { stacked } = this.configs;
-		let scatterData;
-		if (stacked) {
-			const percentage = Object.keys(options.axes).some(
-				(axis) => options.axes[axis].percentage
-			);
-			scatterData = this.model.getStackedData({groups: this.configs.groups, percentage });
-		} else {
-			scatterData = this.model.getDisplayData(this.configs.groups)
-				.filter(
-					(d) => {
-						const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(d);
-						return d[rangeIdentifier] !== undefined && d[rangeIdentifier] !== null;
-					}
-				);
-		}
-
-		// filter out datapoints that aren't part of the zoomed domain
-		scatterData = this.filterBasedOnZoomDomain(scatterData);
+		const domainIdentifier = this.services.cartesianScales.getDomainIdentifier();
 
 		// Update data on dot groups
 		const circles = svg
 			.selectAll("circle.dot")
-			.data(scatterData);
+			.data(
+				this.getScatterData(),
+				(datum) => `${datum[groupMapsTo]}-${datum[domainIdentifier]}`
+			);
 
 		// Remove circles that need to be removed
 		circles.exit().attr("opacity", 0).remove();
@@ -212,28 +231,24 @@ export class Scatter extends Component {
 					d[domainIdentifier],
 					d,
 					filled
-					);
-				}
-			)
-			.classed(
-				"unfilled",
-				(d) => {
-					const domainIdentifier = cartesianScales.getDomainIdentifier(d);
-					return !this.model.getIsFilled(
-						d[groupMapsTo],
-						d[domainIdentifier],
-						d,
-						filled
-					);
-				}
-			)
+				);
+			})
+			.classed("unfilled", (d) => {
+				const domainIdentifier = cartesianScales.getDomainIdentifier(d);
+				return !this.model.getIsFilled(
+					d[groupMapsTo],
+					d[domainIdentifier],
+					d,
+					filled
+				);
+			})
 			.transition(
 				transitions.getTransition("scatter-update-enter", animate)
 			)
 			.attr("cx", getXValue)
 			.attr("cy", getYValue)
 			.attr("r", options.points.radius)
-			.attr("fill", (d) => {
+			.style("fill", (d) => {
 				const domainIdentifier = cartesianScales.getDomainIdentifier(d);
 				if (
 					this.model.getIsFilled(
@@ -250,16 +265,15 @@ export class Scatter extends Component {
 					);
 				}
 			})
-			.attr("fill-opacity", filled ? fillOpacity : 1)
-			.attr("stroke", (d) => {
+			.style("stroke", (d) => {
 				const domainIdentifier = cartesianScales.getDomainIdentifier(d);
 				return this.model.getStrokeColor(
-						d[groupMapsTo],
-						d[domainIdentifier],
-						d
-					);
-				}
-			)
+					d[groupMapsTo],
+					d[domainIdentifier],
+					d
+				);
+			})
+			.attr("fill-opacity", filled ? fillOpacity : 1)
 			.attr("opacity", fadeInOnChartHolderMouseover ? 0 : 1)
 			// a11y
 			.attr("role", Roles.GRAPHICS_SYMBOL)
@@ -282,7 +296,7 @@ export class Scatter extends Component {
 				)
 			)
 			.attr("opacity", 1);
-	}
+	};
 
 	handleChartHolderOnMouseOut = (event: CustomEvent) => {
 		this.parent
@@ -293,7 +307,7 @@ export class Scatter extends Component {
 				)
 			)
 			.attr("opacity", 0);
-	}
+	};
 
 	handleLegendOnHover = (event: CustomEvent) => {
 		const { hoveredElement } = event.detail;
@@ -308,7 +322,7 @@ export class Scatter extends Component {
 			.attr("opacity", (d) =>
 				d[groupMapsTo] !== hoveredElement.datum()["name"] ? 0.3 : 1
 			);
-	}
+	};
 
 	handleLegendMouseOut = (event: CustomEvent) => {
 		this.parent
@@ -319,7 +333,7 @@ export class Scatter extends Component {
 				)
 			)
 			.attr("opacity", 1);
-	}
+	};
 
 	addEventListeners() {
 		const self = this;
@@ -340,13 +354,15 @@ export class Scatter extends Component {
 						})
 					)
 					.style("fill", (d) => {
-						const domainIdentifier = self.services.cartesianScales.getDomainIdentifier(d);
+						const domainIdentifier = self.services.cartesianScales.getDomainIdentifier(
+							d
+						);
 						return self.model.getFillColor(
-								d[groupMapsTo],
-								d[domainIdentifier],
-								d);
-						}
-					)
+							d[groupMapsTo],
+							d[domainIdentifier],
+							d
+						);
+					})
 					.classed("unfilled", false);
 
 				// Show tooltip
@@ -398,12 +414,16 @@ export class Scatter extends Component {
 				const hoveredElement = select(this);
 				hoveredElement.classed("hovered", false);
 
-				if (
-					!self.configs.filled &&
-					hoveredElement.attr("fill-opacity") === "1"
-				) {
-					hoveredElement.classed("unfilled", true);
-					hoveredElement.style("fill", null);
+				if (!self.configs.filled) {
+					const { filled } = self.getOptions().points;
+					const domainIdentifier = self.services.cartesianScales.getDomainIdentifier(datum);
+					hoveredElement.classed("unfilled", !self.model.getIsFilled(
+						datum[groupMapsTo],
+						datum[domainIdentifier],
+						datum,
+						filled
+					))
+					.style("fill", null);
 				}
 
 				// Dispatch mouse event