+++ packages/core/src/configuration.ts
@@ -3,6 +3,7 @@ import {
 	BaseChartOptions,
 	AxisChartOptions,
 	ScatterChartOptions,
+	LollipopChartOptions,
 	LineChartOptions,
 	BarChartOptions,
 	StackedBarChartOptions,
@@ -245,6 +246,11 @@ const scatterChart: ScatterChartOptions = Tools.merge({}, axisChart, {
 } as ScatterChartOptions);
 
 /**
+ * options specific to lollipop charts
+ */
+const lollipopChart: LollipopChartOptions = scatterChart as LollipopChartOptions;
+
+/**
  * options specific to line charts
  */
 const lineChart: LineChartOptions = Tools.merge({}, scatterChart, {
@@ -382,8 +388,10 @@ const radarChart: RadarChartOptions = Tools.merge({}, chart, {
 
 /**
  * options specific to combo charts
-*/
-const comboChart: ComboChartOptions = baseBarChart;
+ */
+const comboChart: ComboChartOptions = Tools.merge({}, baseBarChart, {
+	comboChartTypes: []
+} as ComboChartOptions);
 
 /*
  * options specific to treemap charts
@@ -405,6 +413,7 @@ export const options = {
 	areaChart,
 	stackedAreaChart,
 	scatterChart,
+	lollipopChart,
 	pieChart,
 	donutChart,
 	meterChart,