+++ packages/core/src/interfaces/charts.ts
@@ -1,4 +1,10 @@
-import { GaugeTypes, Statuses, ArrowDirections, Alignments, ChartTypes } from "./enums";
+import {
+	GaugeTypes,
+	Statuses,
+	ArrowDirections,
+	Alignments,
+	ChartTypes
+} from "./enums";
 import {
 	LegendOptions,
 	TooltipOptions,
@@ -171,6 +177,11 @@ export interface ScatterChartOptions extends AxisChartOptions {
 }
 
 /**
+ * options specific to lollipop charts
+ */
+export interface LollipopChartOptions extends ScatterChartOptions {}
+
+/**
  * options specific to bubble charts
  */
 export interface BubbleChartOptions extends AxisChartOptions {
@@ -213,7 +224,7 @@ export interface LineChartOptions extends ScatterChartOptions {
 		| string
 		| {
 				name: string;
-	};
+		  };
 }
 
 /**
@@ -227,7 +238,7 @@ export interface AreaChartOptions extends AxisChartOptions {
 		| string
 		| {
 				name: string;
-		};
+		  };
 }
 
 /**
@@ -241,7 +252,7 @@ export interface StackedAreaChartOptions extends ScatterChartOptions {
 		| string
 		| {
 				name: string;
-		};
+		  };
 }
 
 /**
@@ -321,18 +332,15 @@ export interface RadarChartOptions extends BaseChartOptions {
 }
 
 /**
- * options specific to combo bar charts
+ * options specific to combo charts
  */
 export interface ComboChartOptions extends AxisChartOptions {
-	comboChartTypes?: {
-		[ChartTypes.LINE]?: [],
-		[ChartTypes.STACKED_BAR]?: [],
-		[ChartTypes.GROUPED_BAR]?: [],
-		[ChartTypes.SIMPLE_BAR]?: [],
-		[ChartTypes.SCATTER]?: []
-	};
+	comboChartTypes: Array<{
+		type: ChartTypes | any;
+		options?: object;
+		correspondingDatasets: Array<string>;
+	}>;
 }
-
 
 /*
  * options specific to treemap charts